Add files via upload
 Brand Choice Analysis

We have data on detergent purchased from variety of supermarkets. There are variety of choices available. We observe consumer choices from the characteristics provided by the data. As the utility varies by the characteristics of the choices of the brands and each brand has different price, market share and promotions. The customer chooses brand with highest utility. Therefore, the multinomial conditional logit(MNL) aligns perfectly.